backend/requirements.txt
pandas
scikit-learn
fastapi
uvicorn
python-decouple
PyJWT
joblib